Calculates evaluation measures for a Petri nets and an Event Log

Calculates evaluation measures for a Petri nets and an Event Log

evaluation_all(eventlog, petrinet, initial_marking, final_marking,
  parameters = default_parameters(eventlog), convert = TRUE)

evaluation_precision(eventlog, petrinet, initial_marking, final_marking,
  parameters = default_parameters(eventlog),
  variant = variant_precision_etconformance(), convert = TRUE)

variant_precision_etconformance()

evaluation_fitness(eventlog, petrinet, initial_marking, final_marking,
  parameters = default_parameters(eventlog),
  variant = variant_fitness_token_based(), convert = TRUE)

variant_fitness_token_based()

variant_fitness_alignment_based()

Arguments

eventlog
A bupaR or PM4PY event log.

petrinet
A bupaR or PM4PY Petri net.

initial_marking
A R vector with the place identifiers of the initial marking or a PM4PY marking. By default the initial marking of the bupaR Petri net will be used if available.

final_marking
A R vector with the place identifiers of the final marking or a PM4PY marking.

parameters
PM4PY alignment parameter. By default the activity_key from the bupaR event log is specified using param_activity_key.

convert
TRUE to automatically convert Python objects to their R equivalent. If you pass FALSE you can do manual conversion using the r-py-conversion function.

variant
The evaluation variant to be used.

Value

A list with all available evaluation measures.

Examples

if (pm4py_available()) {
  library(eventdataR)
  data(patients)

  # As Inductive Miner of PM4PY is not life-cycle aware, keep only `complete` events:
  patients_completes <- patients[patients$registration_type == "complete", ]

  # Discover a Petri net
  net <- discovery_inductive(patients_completes)

  # Calculate evaluation measures for event log and Petri net
  evaluation_all(patients_completes,
                 net$petrinet,
                 net$initial_marking,
                 net$final_marking)

}

#> $fitness
#> $fitness$perc_fit_traces
#> [1] 100
#> 
#> $fitness$average_trace_fitness
#> [1] 1
#> 
#> $fitness$log_fitness
#> [1] 1
#> 
#> 
#> $precision
#> [1] 0.3540365
#> 
#> $generalization
#> [1] 0.6246695
#> 
#> $simplicity
#> [1] 0.65
#> 
#> $metricsAverageWeight
#> [1] 0.6571765
#>